services/pipeline-service/requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
pandas>=2.2.0
numpy>=1.26.0
psycopg[binary,pool]>=3.2.0
python-dotenv>=1.0.1
huggingface_hub>=0.25.0
websockets>=14.0
httpx>=0.27.0
